mas cambios

+++ 1.-FRONT/goodvibing_app/src/DeleteComment.js
@@ -1,30 +1,37 @@
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
+import { useHistory, useParams } from "react-router";
 
-function DeleteStory({ comment_id }) {
+function DeleteComment() {
   const token = useSelector((s) => s.user?.token);
+  const { id } = useParams();
+  const dispatch = useDispatch();
+  const history = useHistory();
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    const res = await fetch(
-      `https://localhost:4000/api/comments/${comment_id}`,
-      {
+    if (window.confirm("¿Estás seguro de eliminar tu comentario?")) {
+      const res = await fetch(`https://localhost:4000/api/comments/${id}`, {
         method: "DELETE",
         headers: {
           "Content-Type": "application/json",
           Authorization: "Bearer " + token,
         },
+      });
+
+      const data = await res.json();
+      if (res.ok) {
+        history.push("/getcomments");
+      } else {
+        dispatch({ type: "SET_ERROR", message: data.message });
       }
-    );
-    if (res.ok) {
-      await res.json();
     }
   };
 
   return (
-    <div className="deletecomment" onSubmit={handleSubmit}>
+    <form onSubmit={handleSubmit}>
       <button>Eliminar comentario</button>
-    </div>
+    </form>
   );
 }
 
-export default DeleteStory;
+export default DeleteComment;
